requirements.txt
python==3.9
annotated-types==0.6.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==42.0.5
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
entrypoints==0.4
executing==1.2.0
fastapi==0.109.2
fastjsonschema==2.16.2
filelock==3.8.0
glfw==2.5.9
h11==0.14.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.3.0
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pickleshare==0.7.5
Pillow==9.5.0
pipenv==2022.11.25
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.23.2
psutil==5.9.4
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pygame==2.4.0
pyglfw==0.2.2
Pygments==2.13.0
PyJWT==2.8.0
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.6
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.31.0
requests-oauthlib==2.0.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.36.3
sympy==1.12
terminado==0.17.0
tinycss2==1.2.1
torch==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.9.0
tzdata==2022.6
urllib3==2.2.1
uvicorn==0.27.1
virtualenv==20.16.7
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
